tests: arch: arm: move test arm_hardfault_validation to new ztest API Migrate the testsuite tests/arch/arm/arm_hardfault_validation to the new ztest API.
