+++ tests/arch/arm/arm_hardfault_validation/prj.conf
@@ -1,1 +1,2 @@
 CONFIG_ZTEST=y
+CONFIG_ZTEST_NEW_API=y
